Move host functions to individual files
